+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 071217 - 2140
+
+- Added image support
+- I need some help with exception handling in commands.py, appreciate help if anyone can read the comments there.
+- cant seem to catch the socket timing out exception there.
+- wand and imagemagick are required for images.
+
 ## 071217 - 1130
 
 - Rewrote code to make it easier for users to customize after vice linked to this bot